requirements.txt
beautifulsoup4==4.4.1
python-dateutil==2.5.3
requests==2.10.0
six==1.10.0
SQLAlchemy==1.0.14
SQLAlchemy-Utils==0.32.9
